Some general cleanup of the code as far as spacing is concerned.
Added whitelist functionality to the allowed attributes portion
of the class. Both blacklist and whitelist work, but not in tandem.
If one is selected, the other will not work, and whitelisting checks
first.

Added new setting that will allow the user to force the text into
utf-8 when it isn't already. It's basic and only uses utf8_encode,
but when preparing info for XML, it's a handy setting to be able
to run.

 Changes to be committed:
	modified:   htmLawed.class.php